@@ -1,16 +1,16 @@
 source 'https://rubygems.org'
 git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-ruby '3.1.0'
+# ruby '3.3.4'
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails', branch: 'main'
-gem 'rails', '>= 6.1.7.8'
+gem 'rails', '>= 7.1.3.4'
 group :development, :production do
   # Use mysql as the database for Active Record
   gem 'mysql2', '~> 0.5'
 end
 # Use Puma as the app server
-gem 'puma', '~> 5.6'
+gem 'puma', '~> 6.0'
 # Use SCSS for stylesheets
 
 gem 'sass-rails', '>= 6'
@@ -23,7 +23,7 @@ gem 'turbolinks', '~> 5'
 # Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
 gem 'jbuilder', '~> 2.7'
 # Use Redis adapter to run Action Cable in production
-gem 'redis'
+gem 'redis', '~> 4.0'
 gem 'hiredis'
 # Use Active Model has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
@@ -46,7 +46,7 @@ group :development do
   gem 'web-console', '>= 4.1.0'
   # Display performance information such as SQL time and flame graphs for each request in your browser.
   # Can be configured to work on production as well see: https://github.com/MiniProfiler/rack-mini-profiler/blob/master/README.md
-  gem 'rack-mini-profiler', '~> 2.0'
+  # gem 'rack-mini-profiler', '~> 2.0'
   gem 'listen', '~> 3.3'
   # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
   gem 'spring'
@@ -76,7 +76,7 @@ group :test do
   # Easy installation and use of web drivers to run system tests with browsers
   gem 'webdrivers'
   # sqlite db
-  gem 'sqlite3'
+  gem 'sqlite3', '~> 1.7', '>= 1.7.3'
 
   gem 'rspec'
   gem 'rspec-rails'
@@ -169,14 +169,14 @@ gem 'bootstrap-daterangepicker-rails'
 #
 gem 'stripe'
 gem 'stripe_event'
-gem 'noid-rails'
+gem 'noid-rails', git: 'https://github.com/samvera/noid-rails'
 gem 'clipboard-rails'
 #
 #
 # # Analystics
 gem 'ahoy_matey'
 #
-gem 'tinymce-rails'
+gem 'tinymce-rails', '~> 5.0'
 gem 'open_uri_redirections'
 #
 gem 'ibm_watson'
@@ -228,7 +228,7 @@ gem 'invisible_captcha'
 # # API
 gem 'devise_token_auth'
 gem "paranoia", "~> 2.2"
-gem 'rails_same_site_cookie'
+# gem 'rails_same_site_cookie'
 #
 gem 'docx'
 gem 'rest-client'
@@ -236,14 +236,14 @@ gem 'rest-client'
 gem 'remove_emoji'
 #
 gem 'htmlentities'
-gem 'yomu'
+gem 'yomu', git: 'https://github.com/smntb/yomu', branch: 'v0.1.11'
 gem 'sidekiq-limit_fetch'
 #
 gem 'rqrcode'
 #
 gem 'pry-remote'
 #
-gem 'turnout'
+# gem 'turnout'
 #
 gem 'simple_hubspot'
 gem 'social-share-button'
