@@ -55,7 +55,7 @@ class SearchBuilder < Blacklight::SearchBuilder
     condition_limiter = ["document_type_ss:#{type}", status_condition]
     condition_limiter << "organization_id_is:#{current_organization.id}" unless current_organization.blank?
     solr = CollectionResource.solr_connect
-    collections_raw = solr.post "select?#{URI.encode_www_form({ q: query_collection, defType: 'lucene', fq: condition_limiter, fl: title_field })}"
+    collections_raw = solr.post "select?#{URI.encode_www_form({ q: query_collection, defType: 'lucene', fq: condition_limiter, wt: 'json', fl: title_field })}"
     counter = 0
     fq_filters_inner = ''
     collections_raw['response']['docs'].each do |single_collection|
@@ -308,7 +308,7 @@ class SearchBuilder < Blacklight::SearchBuilder
       query_string = ''
       term = session_solr[session_solr.keys.first]['keyword_searched'].dup
       term = SearchBuilder.remove_illegal_characters(term, 'simple')
-      all_terms = term.gsub(/OR/, '|||').gsub(/AND/, '|||').gsub(/NOT/, '|||').split('|||').map(&:strip)
+      all_terms = term.gsub('OR', '|||').gsub('AND', '|||').gsub('NOT', '|||').split('|||').map(&:strip)
       filed_name = if session_solr[session_solr.keys.first].keys.include?('title_text')
                      'title_text'
                    elsif session_solr[session_solr.keys.first].keys.include?('resource_description')
