@@ -2,5 +2,5 @@
 #
 # Aviary is an audiovisual content publishing platform with sophisticated features for search and permissions controls.
 # Copyright (C) 2019 Audio Visual Preservation Solutions, Inc.
-# class Admin::UserPresenter < UserPresenter
-# end
+class Admin::UserPresenter < UserPresenter
+end
