@@ -6,7 +6,7 @@ Rails.application.configure do
   # In the development environment your application's code is reloaded any time
   # it changes. This slows down response time but is perfect for development
   # since you don't have to restart the web server when you make code changes.
-  config.cache_classes = false
+  config.enable_reloading = true
 
   # Do not eager load code on boot.
   config.eager_load = false
@@ -14,35 +14,38 @@ Rails.application.configure do
   # Show full error reports.
   config.consider_all_requests_local = true
 
+  # Enable server timing
+  config.server_timing = true
+
   # Enable/disable caching. By default caching is disabled.
   # Run rails dev:cache to toggle caching.
-  if Rails.root.join('tmp', 'caching-dev.txt').exist?
+  if Rails.root.join('tmp/caching-dev.txt').exist?
     config.action_controller.perform_caching = true
     config.action_controller.enable_fragment_cache_logging = true
 
-    config.cache_store = :redis_cache_store, { url: 'redis://localhost:6379/1' }
+    config.cache_store = :redis_cache_store, { url: "redis://#{ENV.fetch('REDIS_SERVER', 'localhost')}:6379/1" }
     config.public_file_server.headers = {
       'Cache-Control' => "public, max-age=#{2.days.to_i}"
     }
   else
     config.action_controller.perform_caching = false
-
-    config.cache_store = :redis_cache_store, { url: 'redis://localhost:6379/1' }
+    config.cache_store = :redis_cache_store, { url: "redis://#{ENV.fetch('REDIS_SERVER', 'localhost')}:6379/1" }
   end
   config.session_store :cache_store, key: ENV['SESSION_KEY'] || '_aviary_key', domain: :all
   # Store uploaded files on the local file system (see config/storage.yml for options).
   config.active_storage.service = :local
-
-  # Don't care if the mailer can't send.
-  # Don't care if the mailer can't send.
-  config.action_mailer.raise_delivery_errors = false
   # development default url with subdomain
-  config.action_mailer.default_url_options = { host: ENV['APP_HOST'], port: 3000 }
+  config.action_mailer.default_url_options = { host: ENV.fetch('APP_HOST', nil), port: 3000 }
 
   # development letter opener configuration
   config.action_mailer.delivery_method = :letter_opener
   config.action_mailer.perform_deliveries = true
-  config.action_mailer.asset_host = "http://#{ENV['APP_HOST']}"
+  config.action_mailer.asset_host = "http://#{ENV.fetch('APP_HOST', nil)}"
+
+  # Don't care if the mailer can't send.
+  config.action_mailer.raise_delivery_errors = false
+
+  config.action_mailer.perform_caching = false
 
   # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log
@@ -59,6 +62,9 @@ Rails.application.configure do
   # Highlight code that triggered database queries in logs.
   config.active_record.verbose_query_logs = true
 
+  # Highlight code that enqueued background job in logs.
+  config.active_job.verbose_enqueue_logs = true
+
   # Debug mode disables concatenation and preprocessing of assets.
   # This option may cause significant delays in view rendering with a large
   # number of complex assets.
@@ -73,13 +79,26 @@ Rails.application.configure do
   # Annotate rendered view with file names.
   # config.action_view.annotate_rendered_view_with_filenames = true
 
+  # Uncomment if you wish to allow Action Cable access from any origin.
+  # config.action_cable.disable_request_forgery_protection = true
+
+  # Raise error when a before_action's only/except options reference missing actions
+  config.action_controller.raise_on_missing_callback_actions = true
+
   # Use an evented file watcher to asynchronously detect changes in source code,
   # routes, locales, etc. This feature depends on the listen gem.
   config.file_watcher = ActiveSupport::EventedFileUpdateChecker
-
-  # Uncomment if you wish to allow Action Cable access from any origin.
-  # config.action_cable.disable_request_forgery_protection = true
   config.hosts << '.lvh.me'
 
-  # config.assets.check_precompiled_asset = false
+  config.assets.check_precompiled_asset = false
+  config.paperclip_defaults = { storage: :s3,
+                                s3_protocol: :https,
+                                url: ':s3_alias_url',
+                                s3_host_alias: ENV.fetch('S3_HOST_CDN', nil),
+                                path: '/:class/:attachment/:id_partition/:style/:filename',
+                                s3_credentials: { bucket: ENV.fetch('S3_BUCKET_NAME', nil),
+                                                  access_key_id: ENV.fetch('AWS_ACCESS_KEY_ID', nil),
+                                                  secret_access_key: ENV.fetch('AWS_SECRET_ACCESS_KEY', nil),
+                                                  s3_region: ENV.fetch('S3_REGION', nil),
+                                                  s3_host_name: ENV.fetch('S3_HOST_NAME', nil) } }
 end
